% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{folders}
\alias{folders}
\title{Build folders}
\usage{
folders(...)
}
\arguments{
\item{...}{}
}
\value{
folders
}
\description{
Build folders
}
\examples{
Setup R projects with standard folder outputs:

folders() # creates a data, code, figs, and ouput folder by default
the ... can be used to create additional folders
folders("tables", "text", "tmb")
}
