% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Get the mode of a variable}
\usage{
mode(x)
}
\arguments{
\item{x}{= input variable}
}
\value{
mode
}
\description{
Get the mode of a variable
}
\examples{
df <- data.frame(year = 1970:2019,
               catch = rnorm(50, 10, 2))
mode(df$catch)
}
