% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_tickr.R
\name{scale_x_tickr}
\alias{scale_x_tickr}
\title{Adjust x-axis tick marks and labels}
\usage{
scale_x_tickr(data, var, to = 5, start = NULL, end = NULL, min = NULL)
}
\arguments{
\item{data}{= input dataframe}

\item{var}{= variable of interest e.g., year}

\item{to}{= step increase desired e.g., every 5 years}

\item{start}{= adjust the start value}

\item{end}{= adjust the end vlaue}

\item{min}{= lowest value to label}
}
\value{
scale_x_tickr
}
\description{
Adjust x-axis tick marks and labels
}
