% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sleek.R
\name{theme_sleek}
\alias{theme_sleek}
\title{Set your figure theme consistently}
\usage{
theme_sleek(base_size = 11, base_family = "Times")
}
\arguments{
\item{base_family}{}
}
\value{

}
\description{
Set your figure theme consistently
}
\examples{
theme_sleek(base_size = 11, base_family = "Times")

Other fonts are available, though sans font is
the easiest to implement using the following.

theme_sleek(base_family = "")

Updating font size is accomplished by changing the base_size.

theme_sleek(base_size = 20, base_family = "")

}
